pyproject.toml
[tool.poetry]
name = "beano"
version = "0.1.0"
description = "AI system to brainstorm ideas"
authors = ["Lorenzo Micheli <lorenzo.micheli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.44"
langchain-aws = "^0.2.6"
langchain = "^0.3.7"
python-dotenv = "^1.0.1"
langgraph-checkpoint-sqlite = "^2.0.1"
tavily-python = "^0.5.0"
pydantic = "^2.9.2"
watchdog = "^6.0.0"
duckduckgo-search = "^6.3.7"
beautifulsoup4 = "^4.12.3"
nest-asyncio = "^1.6.0"
langchain-community = "^0.3.8"
streamlit = "^1.40.2"
pyperclip = "^1.9.0"
googlesearch-python = "1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ipykernel = "^6.29.5"
pytest-cov = "^6.0.0"
bandit = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
